requirements.txt
pre-commit==2.15.0
psycopg2-binary==2.9.5
django==3.2
python-dotenv==0.19.2
split_settings==1.0.0
django-debug-toolbar==3.2.4
python-dateutil==2.8.2
gunicorn==20.1.0
elasticsearch==8.6.0
pydantic==1.10.2
pandas==1.5.0
pytest==7.2.0
pytest-asyncio==0.20.3
flake8==6.0.0
black==22.12.0
flake8-isort==6.0.0
more-itertools==9.0.0
redis==4.4.2
apache-airflow==2.5.0
